Tidy unit tests
